erstatt egen kafka-producer med kode som del av initial load-task
