+++ mulighetsrommet-api/src/main/kotlin/no/nav/mulighetsrommet/api/gjennomforing/task/InitialLoadGjennomforinger.kt
@@ -4,8 +4,9 @@ import com.github.kagkarlsson.scheduler.SchedulerClient
 import com.github.kagkarlsson.scheduler.task.helper.OneTimeTask
 import com.github.kagkarlsson.scheduler.task.helper.Tasks
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.Json
+import no.nav.common.kafka.producer.KafkaProducerClient
 import no.nav.mulighetsrommet.api.ApiDatabase
-import no.nav.mulighetsrommet.api.gjennomforing.kafka.SisteTiltaksgjennomforingerV1KafkaProducer
 import no.nav.mulighetsrommet.api.gjennomforing.mapper.TiltaksgjennomforingEksternMapper
 import no.nav.mulighetsrommet.api.gjennomforing.model.GjennomforingDto
 import no.nav.mulighetsrommet.database.utils.DatabaseUtils.paginateFanOut
@@ -14,14 +15,20 @@ import no.nav.mulighetsrommet.model.Tiltakskode
 import no.nav.mulighetsrommet.serializers.UUIDSerializer
 import no.nav.mulighetsrommet.tasks.DbSchedulerKotlinSerializer
 import no.nav.mulighetsrommet.tasks.executeSuspend
+import org.apache.kafka.clients.producer.ProducerRecord
 import org.slf4j.LoggerFactory
 import java.time.Instant
 import java.util.*
 
 class InitialLoadGjennomforinger(
+    private val config: Config,
     private val db: ApiDatabase,
-    private val gjennomforingProducer: SisteTiltaksgjennomforingerV1KafkaProducer,
+    private val kafkaProducerClient: KafkaProducerClient<ByteArray, ByteArray?>,
 ) {
+    data class Config(
+        val topic: String,
+    )
+
     private val logger = LoggerFactory.getLogger(javaClass)
 
     @Serializable
@@ -94,7 +101,7 @@ class InitialLoadGjennomforinger(
             val gjennomforing = queries.gjennomforing.get(id)
             if (gjennomforing == null) {
                 logger.info("Sender tombstone for id $id")
-                gjennomforingProducer.retract(id)
+                retract(id)
             } else {
                 logger.info("Publiserer melding for $id")
                 publish(gjennomforing)
@@ -110,6 +117,22 @@ class InitialLoadGjennomforinger(
 
     private fun publish(dto: GjennomforingDto) {
         val message = TiltaksgjennomforingEksternMapper.toTiltaksgjennomforingV1Dto(dto)
-        gjennomforingProducer.publish(message)
+
+        val record: ProducerRecord<ByteArray, ByteArray?> = ProducerRecord(
+            config.topic,
+            message.id.toString().toByteArray(),
+            Json.encodeToString(message).toByteArray(),
+        )
+
+        kafkaProducerClient.sendSync(record)
+    }
+
+    fun retract(id: UUID) {
+        val record: ProducerRecord<ByteArray, ByteArray?> = ProducerRecord(
+            config.topic,
+            id.toString().toByteArray(),
+            null,
+        )
+        kafkaProducerClient.sendSync(record)
     }
 }
